refactor(kmod): split agnocast_main.c into main, ioctl
